requirements.txt
ply==3.11
google-cloud-pubsub==2.17.1
google-cloud-storage==2.9.0